+++ markwon-core/src/main/java/io/noties/markwon/urlprocessor/UrlProcessor.java
@@ -1,8 +1,0 @@
-package io.noties.markwon.urlprocessor;
-
-import androidx.annotation.NonNull;
-
-public interface UrlProcessor {
-    @NonNull
-    String process(@NonNull String destination);
-}