+++ sample/src/main/java/io/noties/markwon/sample/recycler/RecyclerActivity.java
@@ -34,6 +34,8 @@ import io.noties.markwon.MarkwonVisitor;
 import io.noties.markwon.core.CorePlugin;
 import io.noties.markwon.html.HtmlPlugin;
 import io.noties.markwon.image.ImagesPlugin;
+import io.noties.markwon.image.destination.ImageDestinationProcessor;
+import io.noties.markwon.image.destination.ImageDestinationProcessorRelativeToAbsolute;
 import io.noties.markwon.image.file.FileSchemeHandler;
 import io.noties.markwon.image.network.OkHttpNetworkSchemeHandler;
 import io.noties.markwon.image.svg.SvgMediaDecoder;
@@ -42,8 +44,6 @@ import io.noties.markwon.recycler.SimpleEntry;
 import io.noties.markwon.recycler.table.TableEntry;
 import io.noties.markwon.recycler.table.TableEntryPlugin;
 import io.noties.markwon.sample.R;
-import io.noties.markwon.urlprocessor.UrlProcessor;
-import io.noties.markwon.urlprocessor.UrlProcessorRelativeToAbsolute;
 
 public class RecyclerActivity extends Activity {
 
@@ -100,7 +100,7 @@ public class RecyclerActivity extends Activity {
                 .usePlugin(new AbstractMarkwonPlugin() {
                     @Override
                     public void configureConfiguration(@NonNull MarkwonConfiguration.Builder builder) {
-                        builder.urlProcessor(new UrlProcessorInitialReadme());
+                        builder.imageDestinationProcessor(new ImageDestinationProcessorInitialReadme());
                     }
 
                     @Override
@@ -182,12 +182,12 @@ public class RecyclerActivity extends Activity {
         return out;
     }
 
-    private static class UrlProcessorInitialReadme implements UrlProcessor {
+    private static class ImageDestinationProcessorInitialReadme extends ImageDestinationProcessor {
 
         private static final String GITHUB_BASE = "https://github.com/noties/Markwon/raw/master/";
 
-        private final UrlProcessorRelativeToAbsolute processor
-                = new UrlProcessorRelativeToAbsolute(GITHUB_BASE);
+        private final ImageDestinationProcessorRelativeToAbsolute processor
+                = new ImageDestinationProcessorRelativeToAbsolute(GITHUB_BASE);
 
         @NonNull
         @Override